+++ eslint.config.mjs
@@ -1,0 +1,4 @@
+import config from '@tuomashatakka/eslint-config'
+
+
+export default config