+++ jest.config.js
@@ -3,7 +3,15 @@ module.exports = {
   testEnvironment: 'node',
   roots: ['<rootDir>/src'],
   testMatch: ['**/__tests__/**/*.test.ts'],
+  transform: {
+    '^.+\\.tsx?$': ['ts-jest', {
+      isolatedModules: true,
+      diagnostics: {
+        ignoreCodes: ['2305', '2304'],
+      },
+    }],
+  },
   moduleNameMapper: {
-    '^vscode$': '<rootDir>/src/__mocks__/vscode.ts'
-  }
+    '^vscode$': '<rootDir>/src/__mocks__/vscode.ts',
+  },
 };